Delete ContentPatchSeeder — the replace supersedes it
The row patch existed to reach an existing install *without* replacing the
content database. ContentArtifactInstaller replaces it, so the patch is now a
second delivery path that can only do less: `nz patch emit` cannot express a
table absent from the baseline — it files those under `out_of_scope` and emits
nothing — which is exactly how the thematic layer shipped with five empty tables
on every device that already had the app.

Two paths where one silently under-delivers is worse than one path. Gone:
ContentPatchSeeder and its test, the applyContentPatch step in AppInitializer,
both DI bindings, contentPatchVersion on SettingsRepository and
PreferencesDataStore, the CONTENT_PATCH_VERSION key, and the `patch` entry the
Gradle fetch pulled out of data.lock.json. USER_TABLES moves to the installer,
which still needs it to know when a pre-schemaVersion-23 install has rows that
have not been copied out yet.

I had the trade backwards when I proposed keeping the patch for cheap
corrections, and the docs said so: replacing does *not* cost an extra 54 MB
download. The artifact is fetched at build time and registered as a generated
assets source root, so every APK already carries the artifact its data.lock.json
pins — the device gets it with the app update either way. The patch was an
*additional* ~2.7 MB asset on top, so deleting it makes the APK slightly
smaller. What the replace actually costs is a one-time ~176 MB copy, disk to
disk with no network, on the first launch after a content change — the same copy
a fresh install has always paid. SUBSYSTEMS.md and the retirement ledger are
corrected.

`nz patch emit` / `nz patch verify` stay in arshad-shah/nimaz-data. Nothing in
this repository reads a patch now, but the emitter is the fallback if a delivery
path that does not rewrite the whole file is ever wanted again.
